+++ src/constants/feed.ts
@@ -1,0 +1,3 @@
+const PAGE_SIZE = 15;
+
+export default PAGE_SIZE;
